// pwa/src/app/models/card.ts
import { ChildrenNumberPropertiesSum } from './children-number-properties-sum';
import { CardProperties } from './card-properties';
import { ShareSettings } from './share-settings';
import { Source } from './source';
import { Tag } from './tag';
import { Owner } from './owner';
import { CardType } from './card-type';
import { ExternalLink } from './external-link';
import { CardState } from './card-state';

export interface Card {
  id: number;
  uid: string;
  created: Date;
  updated: Date;
  archived: boolean;
  title: string;
  asap: boolean;
  due_date: null;
  sort_order: number;
  fifo_order: number | null;
  state: CardState;
  condition: number;
  expires_later: boolean;
  parents_count: number;
  children_count: number;
  children_done: number;
  has_blocked_children: boolean;
  goals_total: number;
  goals_done: number;
  time_spent_sum: number;
  time_blocked_sum: number;
  children_number_properties_sum: ChildrenNumberPropertiesSum | null;
  parent_checklist_ids: null;
  parents_ids: number[] | null;
  children_ids: number[] | null;
  blocking_card: boolean;
  blocked: boolean;
  size: number | null;
  size_unit: string | null;
  size_text: string | null;
  due_date_time_present: boolean;
  board_id: number;
  column_id: number;
  lane_id: number;
  owner_id: number;
  type_id: number;
  version: number;
  updater_id: number;
  completed_on_time: null;
  completed_at: null;
  last_moved_at: Date;
  lane_changed_at: Date;
  column_changed_at: Date;
  first_moved_to_in_progress_at: Date | null;
  last_moved_to_done_at: Date | null;
  sprint_id: null;
  external_id: null;
  comments_total: number;
  comment_last_added_at: Date | null;
  properties: CardProperties | null;
  counters_recalculated_at: Date;
  service_id: null;
  sd_new_comment: boolean;
  public: boolean;
  share_id: null | string;
  share_settings: ShareSettings | null;
  external_user_emails: null;
  calculated_planned_start: null;
  calculated_planned_end: null;
  planned_start: null;
  planned_end: null;
  description_filled: boolean;
  estimate_workload: number;
  tag_ids: number[] | null;
  locked: null;
  source: Source | null;
  description: string;
  tags?: Tag[];
  owner: Owner;
  members?: Owner[];
  type: CardType;
  external_links?: ExternalLink[];
}